Lab_07 finished

+++ C_Labs/lab_07_01_04/src/tools.c
@@ -41,7 +41,7 @@ double get_avg(const int *pcur, const int *pend)
         ++count;
     }
 
-    return summ / count;
+    return (double) summ / count;
 }
